+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 1.2.0 / 2018-05-3
 ===================
 - Use first non-empty address line in screen reader only text for Map and Direction and Services Available links
+- Allow fonts from hotjar.com in Content Security Policy
 
 1.1.0 / 2018-04-17
 ===================